requirements.txt
Django
psycopg2
djangorestframework
django-simple-history
pillow
python-dateutil